Refactor k3s utils for reuse
